Split prompt composer recording surfaces
