+++ scripts/design-token-baseline.txt
@@ -4,7 +4,7 @@
 # lower numbers are better — re-baseline after migrating a screen.
 app/src/main/java/com/pocketshell/app/composer/MarkdownText.kt 4
 app/src/main/java/com/pocketshell/app/composer/PromptComposerQueueBanners.kt 2
-app/src/main/java/com/pocketshell/app/composer/PromptComposerSheet.kt 20
+app/src/main/java/com/pocketshell/app/composer/PromptComposerSheet.kt 13
 app/src/main/java/com/pocketshell/app/conversation/ConversationMessageTurn.kt 1
 app/src/main/java/com/pocketshell/app/conversation/ConversationTextSection.kt 3
 app/src/main/java/com/pocketshell/app/crash/CrashReportsScreen.kt 2
@@ -15,7 +15,7 @@ app/src/main/java/com/pocketshell/app/hosts/HostListScreen.kt 1
 app/src/main/java/com/pocketshell/app/hosts/QrScannerScreen.kt 1
 app/src/main/java/com/pocketshell/app/hosts/SshKeysScreen.kt 3
 app/src/main/java/com/pocketshell/app/projects/FolderListScreen.kt 19
-app/src/main/java/com/pocketshell/app/projects/RepoBrowserScreen.kt 4
+app/src/main/java/com/pocketshell/app/projects/RepoBrowserScreen.kt 2
 app/src/main/java/com/pocketshell/app/projects/WatchedFoldersChipRow.kt 1
 app/src/main/java/com/pocketshell/app/projects/WatchedFoldersScreen.kt 2
 app/src/main/java/com/pocketshell/app/session/ConversationSyncStatusUi.kt 1
@@ -25,4 +25,4 @@ app/src/main/java/com/pocketshell/app/tmux/TmuxSessionConnectionChrome.kt 1
 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionOverlays.kt 7
 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt 4
 app/src/main/java/com/pocketshell/app/usage/UsageResetBanner.kt 2
-app/src/main/java/com/pocketshell/app/usage/UsageScreen.kt 3
+app/src/main/java/com/pocketshell/app/usage/UsageScreen.kt 1
